% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmojiDictionary.R
\docType{data}
\name{EmojiDictionary}
\alias{EmojiDictionary}
\title{Emoji Dictionary}
\format{A .csv dataframe}
\usage{
Used for replacing misrepresented Emojis with respective textual description or unicode
}
\description{
a dictionary of Emojis taken from \link{https://unicode.org/emoji/charts-11.0/full-emoji-list.html} with manually
added WhatsApp specific Emojis
}
\keyword{Emojis}
\keyword{datasets,}
