% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.R
\docType{data}
\name{Example}
\alias{Example}
\title{Example Dataset}
\format{A .txt dataframe}
\usage{
showcasing the functionality of the WhatsApp Parser
}
\description{
a fictitious dataset showcasing the functionality of the WhatsApp Parser
}
\keyword{Textfile}
\keyword{WhatsApp}
\keyword{datasets,}
