% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smilies.R
\docType{data}
\name{Smileys}
\alias{Smileys}
\title{Smiley Dictionary}
\format{A .csv dataframe}
\usage{
Used for Extraction of Smileys in the WhatsAppParser
}
\description{
a dictionary of ASCII Smileys taken from \link{https://de.wiktionary.org/wiki/Verzeichnis:International/Smileys} with manually
added common Smileys
}
\keyword{Smileys}
\keyword{datasets,}
